1. now is working when timestep=1, cfg=1; 2. cfg > 1 is not working, will try to fix in the next commit;

+++ Sana/models/sana_multi_scale.py
@@ -261,7 +261,7 @@ class SanaMS(Sana):
         ## run original forward pass
         out = self.forward_raw(
             x = x.to(self.dtype),
-            timestep = timesteps.to(self.dtype),
+            timestep = timesteps.to(torch.float32),
             y = context.to(self.dtype),
             **kwargs
         )
